move the file functions to the library
