+++ lib.h
@@ -32,11 +32,6 @@
 #include <string.h>
 #include <sys/types.h>
 
-#undef MAX
-#undef MIN
-#define MAX(a, b) ((a) < (b) ? (b) : (a))
-#define MIN(a, b) ((a) < (b) ? (a) : (b))
-
 /* Use this with care.  If the compiler can't see the array
  * definition, it obviously won't produce a correct result. */
 #define mutt_array_size(x) (sizeof(x) / sizeof((x)[0]))
@@ -61,38 +56,9 @@ extern int debuglevel_cmdline;
 #define S_ERR 127
 #define S_BKG 126
 
-/* Flags for mutt_read_line() */
-#define MUTT_CONT (1 << 0) /**< \-continuation */
-#define MUTT_EOL  (1 << 1) /**< don't strip `\n` / `\r\n` */
-
-/* The actual library functions. */
-
-FILE *safe_fopen(const char *path, const char *mode);
-
-char *mutt_concatn_path(char *dst, size_t dstlen, const char *dir,
-                        size_t dirlen, const char *fname, size_t fnamelen);
-char *mutt_concat_path(char *d, const char *dir, const char *fname, size_t l);
-char *mutt_read_line(char *s, size_t *size, FILE *fp, int *line, int flags);
-
-const char *mutt_basename(const char *f);
-
-int mutt_copy_stream(FILE *fin, FILE *fout);
-int mutt_copy_bytes(FILE *in, FILE *out, size_t size);
-int mutt_rx_sanitize_string(char *dest, size_t destlen, const char *src);
 int safe_asprintf(char **, const char *, ...);
-int safe_open(const char *path, int flags);
-int safe_rename(const char *src, const char *target);
-int safe_symlink(const char *oldpath, const char *newpath);
-int safe_fclose(FILE **f);
-int safe_fsync_close(FILE **f);
-int mutt_rmtree(const char *path);
-int mutt_mkdir(const char *path, mode_t mode);
-
-size_t mutt_quote_filename(char *d, size_t l, const char *f);
 
 void mutt_nocurses_error(const char *, ...);
-void mutt_sanitize_filename(char *f, short slash);
-void mutt_unlink(const char *s);
 int mutt_inbox_cmp(const char *a, const char *b);
 
 const char *mutt_strsysexit(int e);
