+++ po/POTFILES.in
@@ -63,6 +63,7 @@ lib/lib_base64.c
 lib/lib_date.c
 lib/lib_debug.c
 lib/lib_exit.c
+lib/lib_file.c
 lib/lib_md5.c
 lib/lib_memory.c
 lib/lib_message.c
